fix: updated tests + included unique check for wlh_id
